+++ geocoding_android/android/src/main/java/com/baseflow/geocoding/Geocoding.java
@@ -4,8 +4,10 @@ import androidx.annotation.Nullable;
 import android.content.Context;
 import android.location.Address;
 import android.location.Geocoder;
-
+import android.os.Build;
+import androidx.annotation.RequiresApi;
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
 
@@ -34,10 +36,27 @@ class Geocoding {
      * @return a list of Address objects. Returns null or empty list if no matches were found or there is no backend service available.
      * @throws java.io.IOException if the network is unavailable or any other I/O problem occurs.
      */
+    @SuppressWarnings("deprecation")
     List<Address> placemarkFromAddress(String address) throws IOException {
-
         final Geocoder geocoder = createGeocoder(androidContext, locale);
-        return geocoder.getFromLocationName(address, 5);
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            return getAddressesWithGeocodeListener(geocoder, address, 5);
+        } else {
+            return geocoder.getFromLocationName(address, 5);
+        }
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.TIRAMISU)
+    private List<Address> getAddressesWithGeocodeListener(Geocoder geocoder, String address, int maxResults) {
+        List<Address> addresses = new ArrayList<>();
+        geocoder.getFromLocationName(address, maxResults, new Geocoder.GeocodeListener() {
+            @Override
+            public void onGeocode(List<Address> geocodedAddresses) {
+                addresses.addAll(geocodedAddresses);
+            }
+        });
+        return addresses;
     }
 
     /**
@@ -48,12 +67,29 @@ class Geocoding {
      * @return a list of Address objects. Returns null or empty list if no matches were found or there is no backend service available.
      * @throws IOException if the network is unavailable or any other I/O problem occurs.
      */
+    @SuppressWarnings("deprecation")
     List<Address> placemarkFromCoordinates(
             double latitude,
             double longitude
     ) throws IOException {
         final Geocoder geocoder = createGeocoder(androidContext, locale);
-        return geocoder.getFromLocation(latitude, longitude, 5);
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            return getLocationWithGeocodeListener(geocoder, latitude, longitude, 5);
+        } else {
+            return geocoder.getFromLocation(latitude, longitude, 5);
+        }
+    }
+
+    @RequiresApi(api = Build.VERSION_CODES.TIRAMISU)
+    private List<Address> getLocationWithGeocodeListener(Geocoder geocoder, double latitude, double longitude, int maxResults) {
+        List<Address> addresses = new ArrayList<>();
+        geocoder.getFromLocation(latitude, longitude, maxResults, new Geocoder.GeocodeListener() {
+            @Override
+            public void onGeocode(List<Address> geocodedAddresses) {
+                addresses.addAll(geocodedAddresses);
+            }
+        });
+        return addresses;
     }
 
     private static Geocoder createGeocoder(Context androidContext, @Nullable Locale locale) {